A web request was made to a URL where all the detailed information of a product is found. And it unfolded in HTML
A web request was made to a URL where the preloaded comments and score of the product are found. And it is shown in HTML

The necessary graphic controls were carried out to be able to make a new comment and punctuation. The sending of data will be fictitious, that is, the sending of the same will not be implemented